little fixes on variable definitions

+++ BreathBetter/app/src/main/java/com/example/mborper/breathbetter/api/Measurement.java
@@ -11,59 +11,59 @@ import java.io.Serializable;
  */
 public class Measurement implements Serializable {
 
-    private double ppm;
-    private double temperature;
+    private int ppm;
+    private int temperature;
     private double latitude;
     private double longitude;
 
     /**
      * Returns the PPM (parts per million) value of the measurement.
      * <p>
-     *      getPpm() ---> Entero:ppm
+     *      getPpm() ---> Natural:ppm
      *
      * @return ppm value as a double.
      */
-    public double getPpm() {
+    public int getPpm() {
         return ppm;
     }
 
     /**
      * Sets the PPM (parts per million) value for the measurement.
      * <p>
-     *      Entero:Ppm ---> setPpm()
+     *      Natural:Ppm ---> setPpm()
      *
      * @param ppm The PPM value to be set.
      */
-    public void setPpm(double ppm) {
+    public void setPpm(int ppm) {
         this.ppm = ppm;
     }
 
     /**
      * Returns the temperature value of the measurement.
      *<p>
-     *      getTemperature() ---> Entero:temperature
+     *      getTemperature() ---> Natural:temperature
      *
      * @return temperature as a double.
      */
-    public double getTemperature() {
+    public int getTemperature() {
         return temperature;
     }
 
     /**
      * Sets the temperature value for the measurement.
      *<p>
-     *      Entero:temperature ---> setTemperature()
+     *      Natural:temperature ---> setTemperature()
      *
      * @param temperature The temperature value to be set.
      */
-    public void setTemperature(double temperature) {
+    public void setTemperature(int temperature) {
         this.temperature = temperature;
     }
 
     /**
      * Returns the latitude where the measurement was taken.
      *<p>
-     *      getLatitude() ---> Entero:latitude
+     *      getLatitude() ---> Real:latitude
      *
      * @return latitude as a double.
      */
@@ -74,7 +74,7 @@ public class Measurement implements Serializable {
     /**
      * Sets the latitude value for the measurement.
      *<p>
-     *      Entero:latitude ---> setLatitude()
+     *      Real:latitude ---> setLatitude()
      *
      * @param latitude The latitude value to be set.
      */
@@ -85,7 +85,7 @@ public class Measurement implements Serializable {
     /**
      * Returns the longitude where the measurement was taken.
      *<p>
-     *      getLongitude() ---> Entero:longitude
+     *      getLongitude() ---> Real:longitude
      *
      * @return longitude as a double.
      */
@@ -96,7 +96,7 @@ public class Measurement implements Serializable {
     /**
      * Sets the longitude value for the measurement.
      *<p>
-     *      Entero:longitude ---> setLongitude()
+     *      Real:longitude ---> setLongitude()
      *
      * @param longitude The longitude value to be set.
      */
